Refactor interpolation methods and update imports
- Reintroduced the `interpolate_reciprocal_path` function in the symbolics module, enhancing the handling of reciprocal lattice paths.
- Removed the previous definition of `interpolate_reciprocal_path` from the geometries module to streamline functionality.
- Updated import statements across multiple files to ensure consistent access to the new interpolation method.
- Improved code organization and readability by consolidating related imports and removing unused references.
